fix : Post, Channel 폴더 위치 변경
