+++ next-app/src/components/views/Channel/ChannelContainer.tsx
