+++ next-app/src/components/views/Channel/index.tsx
