+++ next-app/src/components/views/Post/Card.style.tsx
@@ -1,7 +1,6 @@
 import Image from 'next/image'
 import styled from 'styled-components'
 import { colors } from 'styles/colors'
-import { getTypographyStyles } from 'styles/fonts'
 
 export const Container = styled.div`
   width: 100%;
@@ -26,23 +25,6 @@ export const BodyWrapper = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
-`
-
-export const Title = styled.h2`
-  ${getTypographyStyles('Headline3_B')}
-
-  color: ${colors.gray800};
-`
-
-export const Contents = styled.p`
-  ${getTypographyStyles('Body1_M')}
-
-  color: ${colors.gray800};
-  white-space: normal;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
 `
 
 export const ImageContainer = styled.div`
