+++ next-app/src/components/views/Post/index.ts
