+++ next-app/src/pages/myPost/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next'
-import Post from 'components/mypage/post'
+import Post from 'components/views/Post'
 import Layout from 'components/common/Layout'
 
 const MyChannel: NextPage = () => {
